move BVH to own package and introduce Intersectable interface (suggesting use for e.g. octree later)
